Locked files don't get deleted with python 3.3 and later.
As per the python documentation:
Changed in version 2.7.3: Since lots of code check for sys.platform ==
'linux2', and there is no essential change between Linux 2.x and 3.x,
sys.platform is always set to 'linux2', even on Linux 3.x. In Python 3.3
and later, the value will always be set to 'linux', so it is recommended
to always use the startswith idiom presented above.

https://docs.python.org/2/library/sys.html#sys.platform